% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axt.R
\name{read_axt}
\alias{read_axt}
\title{axt format}
\usage{
read_axt(file)
}
\arguments{
\item{file}{path.}
}
\description{
\itemize{
\item start: ``The first base is numbered 1.''
\item end: ``The end base is included.''
\item strand: ``If the strand value is "-", the values of the aligning organism's start and end
fields are relative to the reverse-complemented coordinates of its chromosome.''
}
}
\seealso{
\url{https://genome.ucsc.edu/goldenPath/help/axt.html}
}
