% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{read_chain}
\alias{read_chain}
\title{chain format}
\usage{
read_chain(file)
}
\arguments{
\item{file}{path.}
}
\description{
The positions are represented as zero-based half-open intervals.
e.g., the first 100 bases is represented as \{start: 0, end: 100\}
}
\seealso{
\url{https://genome.ucsc.edu/goldenPath/help/chain.html}
}
