% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp.R
\name{decipher_otp}
\alias{decipher_otp}
\title{Decipher using a One-Time Pad.}
\usage{
decipher_otp(cipher_str, otp_str,
  alphanum_str = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789")
}
\arguments{
\item{cipher_str}{A cipher created using the one-time pad (string).}

\item{otp_str}{One-time pad (string)}

\item{alphanum_str}{Optional. The alphanumeric string to be used creating the message ciphers.}
}
\description{
Decipher using a One-Time Pad.
}
\examples{
cipher_str <- "VUD39"
otp_str <- "SK0QJYPGXMU1H4BER89OCZTW6VNFA5L37DI2"
decipher(cipher_str, otp_str)
# BILLY
}
