% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{fy_shuffle}
\alias{fy_shuffle}
\title{Create a Fisher–Yates Shuffle.}
\usage{
fy_shuffle(x, key_seeds = character())
}
\arguments{
\item{x}{A vector of values to be shuffled.}

\item{key_seeds}{Optional. A vector of numeric values between 0 and 1.
Used to create a repeatable shuffle.}
}
\description{
The Fisher–Yates shuffle is an algorithm for generating a random permutation of
a finite sequence—in plain terms, the algorithm shuffles the sequence. The algorithm
effectively puts all the elements into a hat; it continually determines the next
element by randomly drawing an element from the hat until no elements remain.
The algorithm produces an unbiased permutation: every permutation is equally likely.
}
\examples{
key_seeds <- fy_shuffle(seq(from = 0, to = 1, by = .001))
alphanum <- strsplit(alphanum_str, "")[[1]]
alphanumkey <- fy_shuffle(alphanum, key_seeds)
}
