\name{GenerateData2}
\alias{GenerateData2}
\title{Convert catch-at-age to cohort-at-age}
\description{
  Convert a matrix of year x age into a matrix of cohort x age}
\usage{GenerateData2(max.age = 10, nb.of.cohort = 20, ...)}
\arguments{
  \item{max.age}{max age}
  \item{nb.of.cohort}{number of cohorts}
  \item{...}{other parameters}
  }
\references{NONE}
\keyword{misc}
