% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp.R
\name{get_otp}
\alias{get_otp}
\title{Create a One-Time Pad.}
\usage{
get_otp(alphanum_str = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789")
}
\arguments{
\item{alphanum_str}{Optional. The alphanumeric string to be used creating the one-time pad.}
}
\description{
In cryptography, the one-time pad (OTP) is an encryption technique
that cannot be cracked, but requires the use of a one-time pre-shared
key the same size as, or longer than, the message being sent. In this
technique, a plaintext is paired with a random secret key (also referred
to as a one-time pad). Then, each bit or character of the plaintext is
encrypted by combining it with the corresponding bit or character from
the pad using modular addition. If the key is truly random, is at least
as long as the plaintext, is never reused in whole or in part, and is
kept completely secret, then the resulting ciphertext will be impossible
to decrypt or break ("One-Time Pad (OTP)". Cryptomuseum.com. Retrieved 2014-03-17.).
}
\examples{
get_otp()
}
