% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp.R
\name{get_otp_ciphers}
\alias{get_otp_ciphers}
\title{Create a data.table of message ciphers using a One-Time Pad.}
\usage{
get_otp_ciphers(message_vct, otp_str,
  alphanum_str = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789")
}
\arguments{
\item{message_vct}{A character vector of strings to be ciphered.}

\item{otp_str}{One-time pad (string)}

\item{alphanum_str}{Optional. The alphanumeric string to be used creating the message ciphers.}
}
\description{
Create a data.table of message ciphers using a One-Time Pad.
}
\examples{
names <- c("Billy", "Bob", "Thornton")
otp_ciphers <- get_otp_ciphers(names, get_otp())
otp_ciphers
attr(otp_ciphers, "otp_str")
attr(otp_ciphers, "alphanum_str")
}
