% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{read_net}
\alias{read_net}
\title{net format}
\usage{
read_net(file)
}
\arguments{
\item{file}{paths.}
}
\description{
\itemize{
\item ali: ``Number of bases in alignments in chain.''
}
}
\seealso{
\url{https://genome.ucsc.edu/goldenPath/help/net.html}
}
