% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf.R
\name{as_paf}
\alias{as_paf}
\alias{read_paf}
\alias{write_paf}
\title{Pairwise mApping Format (PAF)}
\usage{
as_paf(x)

read_paf(file)

write_paf(x, file)
}
\arguments{
\item{x}{data.frame.}

\item{file}{path.}
}
\description{
The spec is in minimap2 manual:
\url{https://lh3.github.io/minimap2/minimap2.html}.
}
