% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dot.R
\name{plot_dot}
\alias{plot_dot}
\alias{facet_grid_chr}
\alias{theme_dotplot}
\title{Dot plot}
\usage{
plot_dot(data, linewidth = 1, breaks = 1e+07)

facet_grid_chr()

theme_dotplot(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/24,
  base_rect_size = base_size/24
)
}
\arguments{
\item{data}{data.frame in PAF format.}

\item{linewidth}{numeric.}

\item{breaks}{numeric width.}

\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
Dot plot
}
