% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safr-package.R
\docType{package}
\name{safr-package}
\alias{safr}
\alias{safr-package}
\title{safr: Summarize and Visualize Sequence Alignment Format Data}
\description{
Reads various formats of sequence alignment/mapping data to summmarize and visualize them.
}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavywatal@gmail.com} (\href{https://orcid.org/0000-0002-4718-7605}{ORCID})

}
\keyword{internal}
