% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singmaf.R
\name{read_singmaf}
\alias{read_singmaf}
\title{Single coverage pairwise alignment in MAF format}
\usage{
read_singmaf(file)
}
\arguments{
\item{file}{paths.}
}
\description{
\itemize{
\item start: ``This is a zero-based number.''
\item size: ``This number is equal to the number of non-dash characters.''
\item strand: ``If the field is "-" then this is the start relative to
the reverse-complemented source sequence.''
}
}
\seealso{
\url{https://genome.ucsc.edu/FAQ/FAQformat.html#format5}
}
