\name{which.cohort}
\alias{which.cohort}
\title{A function that counts and numbers cohorts given a catch at age matrix}
\description{Number the cohort in a matrix}
\usage{which.cohort(mat)}
\arguments{
  \item{mat}{a matrix of data}
}
\examples{
mat <- matrix(NA, nrow = 5, ncol = 7)
which.cohort(mat)
}
\keyword{misc}
