% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safr-package.R
\docType{package}
\name{safr-package}
\alias{safr}
\alias{safr-package}
\title{safr: Summarize and Visualize Sequence Alignment Format Data}
\description{
Reads various formats of sequence alignment/mapping data to summarize and visualize them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heavywatal/safr}
  \item Report bugs at \url{https://github.com/heavywatal/safr/issues}
}

}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavywatal@gmail.com} (\href{https://orcid.org/0000-0002-4718-7605}{ORCID})

}
\keyword{internal}
